Parameter Make.1-Seed

type 'a t

The principle type, which can be folded over using fold_right.

val fold_right : f:('a -> 'b -> 'b) -> 'a t -> init:'b -> 'b

fold_right ~f t ~init combines the elements of t using f from the right.