Module type Alg_structs__Semigroup.Seed1

As Seed but for parameteric types of one variable

type 'a t

val op : 'a t -> 'a t -> 'a t

op x y is an associative operation over all elements x and y of type t